Add framework for header and the current day
The framework for the header and current day will stay the same
regardless of location and information. This will be hardcoded into the
html document to simplify the JS script. The framework should display
the current temperature, the high and low for the day, and the 24 hour
timeline starting from the current time.