Add sunrise and sunsets to the forecast row
The sunrise and sunset events need to be in the forecast row. These are
special events that require a custom object that contains information
that the hourly processing function can work with. This includes the
event epoch time. This allows the sort function to place the events in
the correct order. The time format also needs to be longer than just the
hour. A new formatter was created to add minutes to the event time.