Add functions to create div and format sun events
The code had repeated blocks creating divs. The div creation functions
removes the repetition and centralizes the key code into one function.
The function to format the sunrise and sunset is to make the creation of
that time stamp consistent, regardless of where the time stamp appears
in the app. The functions were implemented in three areas to test if
they performed exactly the same as the old code.